requirements.txt
jira>=3.8.0
google-generativeai>=0.8.0
